+++ src/Octokit.Webhooks/Models/CheckSuiteEvent/CheckSuite.cs
@@ -9,7 +9,7 @@
     public sealed record CheckSuite
     {
         [JsonPropertyName("id")]
-        public int Id { get; init; }
+        public long Id { get; init; }
 
         [JsonPropertyName("node_id")]
         public string? NodeId { get; init; }
@@ -48,7 +48,7 @@
         public string UpdatedAt { get; init; } = null!;
 
         [JsonPropertyName("latest_check_runs_count")]
-        public int LatestCheckRunsCount { get; init; }
+        public long LatestCheckRunsCount { get; init; }
 
         [JsonPropertyName("check_runs_url")]
         public string CheckRunsUrl { get; init; } = null!;
